refactor: consolidate orphan top-level files
Move global_settings_builder to core/widgets/, merge constants.dart
into core/global/constants.dart, move theme.dart to core/, and move
keys.dart to features/login/view/test_keys.dart.
