--- lib/constants.dart
@@ -1,9 +1,0 @@
-import 'package:device_info_plus/device_info_plus.dart';
-import 'package:package_info_plus/package_info_plus.dart';
-
-// Globally accessible variables which are definitely initialized after main().
-late final PackageInfo packageInfo;
-late final AndroidDeviceInfo? androidInfo;
-late final IosDeviceInfo? iosInfo;
-
-const latestSupportedApiVersion = 3;
