+++ lib/features/document_search/view/sliver_search_bar.dart
@@ -5,7 +5,7 @@ import 'package:paperless_mobile/core/database/hive/hive_config.dart';
 import 'package:paperless_mobile/core/database/tables/local_user_account.dart';
 import 'package:paperless_mobile/features/document_search/view/document_search_bar.dart';
 import 'package:paperless_mobile/features/settings/view/manage_accounts_page.dart';
-import 'package:paperless_mobile/features/settings/view/widgets/global_settings_builder.dart';
+import 'package:paperless_mobile/core/widgets/global_settings_builder.dart';
 import 'package:paperless_mobile/features/settings/view/widgets/user_avatar.dart';
 import 'package:provider/provider.dart';
