+++ lib/features/settings/view/widgets/default_share_file_type_setting.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:paperless_mobile/core/model/file_download_type.dart';
-import 'package:paperless_mobile/features/settings/view/widgets/global_settings_builder.dart';
+import 'package:paperless_mobile/core/widgets/global_settings_builder.dart';
 import 'package:paperless_mobile/features/settings/view/widgets/radio_settings_dialog.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
