+++ lib/features/settings/view/widgets/skip_document_prepraration_on_share_setting.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:paperless_mobile/features/settings/view/widgets/global_settings_builder.dart';
+import 'package:paperless_mobile/core/widgets/global_settings_builder.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
 
 class SkipDocumentPreprationOnShareSetting extends StatelessWidget {
