+++ lib/routing/routes/app_logs_route.dart
@@ -4,7 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:paperless_mobile/features/logging/cubit/app_logs_cubit.dart';
 import 'package:paperless_mobile/features/logging/view/app_logs_page.dart';
 import 'package:paperless_mobile/routing/navigation_keys.dart';
-import 'package:paperless_mobile/theme.dart';
+import 'package:paperless_mobile/core/theme.dart';
 
 part 'app_logs_route.g.dart';
