+++ lib/routing/routes/settings_route.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:go_router/go_router.dart';
 import 'package:paperless_mobile/features/settings/view/settings_page.dart';
 import 'package:paperless_mobile/routing/navigation_keys.dart';
-import 'package:paperless_mobile/theme.dart';
+import 'package:paperless_mobile/core/theme.dart';
 
 import 'shells/authenticated_route.dart';
